+++ app/src/validators/area.validator.js
@@ -21,7 +21,7 @@ class AreaValidator {
         logger.debug('Validating body for create area');
         ctx.checkBody('name').notEmpty().len(1, 100);
         ctx.checkBody('application').optional().check((application) => AreaValidator.notEmptyString(application), 'can not be empty');
-        ctx.checkBody('geostore').optional().isHexadecimal();
+        ctx.checkBody('geostore').optional();
         ctx.checkBody('wdpaid').optional().isInt().toInt();
         ctx.checkBody('datasets').optional().isJSON();
         ctx.checkBody('iso').optional().check((iso) => AreaValidator.isObject(iso), 'must be an object');
@@ -40,7 +40,7 @@ class AreaValidator {
         logger.debug('Validating body for update area');
         ctx.checkBody('name').optional().len(2, 100);
         ctx.checkBody('application').optional().check((application) => AreaValidator.notEmptyString(application), 'can not be empty');
-        ctx.checkBody('geostore').optional().isHexadecimal();
+        ctx.checkBody('geostore').optional();
         ctx.checkBody('wdpaid').optional().isInt();
         ctx.checkBody('datasets').optional().isJSON();
         ctx.checkBody('iso').optional().check((iso) => AreaValidator.isObject(iso), 'must be an object');